+++ pwm.h
@@ -2,7 +2,7 @@
 #define PWM_H_
 
 void SetDutyCycle(unsigned int DutyCycle, unsigned int Frequency, int dir, int wheelSel);
-void SetDutyCycleServo(unsigned int DutyCycle);
+void SetDutyCycleServo(double DutyCycle);
 void InitPWM(void);
 void PWM_ISR(void);
